requirements.txt
Werkzeug == 0.16.1
connexion == 1.1.15
certifi >= 2019.6.16
urllib3 >= 1.25.3
python-dateutil >= 2.6.1
setuptools >= 21.0.0
transitions == 0.6.4
pypubsub == 4.0.0
rwlock == 0.0.7
wsgiserver == 1.3
autologging >= 1.2.1    
PyYAML >= 5.1.0
six >= 1.12.0
gevent == 1.4.0
bottle == 0.12.16
flask == 1.0.2
csvvalidator >= 1.2